pyproject.toml
[project]
name = "programming-team-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3.26",
    "langchain-groq>=0.3.4",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
    "setuptools>=80.9.0",
    "streamlit>=1.46.1",
]
